Replaced frames and atoms by content in ligand::evaluate

+++ src/main.cpp
@@ -222,7 +222,7 @@ int main(int argc, char* argv[])
 		vector<float> s0(solution_size), s1(solution_size), s2(solution_size);
 		for (size_t i = 0; i < num_mc_tasks; ++i)
 		{
-			tp.push_back(packaged_task<int()>(bind(&ligand::bfgs, cref(lig), s0.data(), s1.data(), s2.data(), cref(sf), cref(rec), rng(), num_generations, i, num_mc_tasks)));
+			tp.push_back(packaged_task<int()>(bind(&ligand::bfgs, cref(lig), s0.data(), s1.data(), s2.data(), cref(sf), cref(rec), rng(), num_generations, static_cast<unsigned int>(i), static_cast<unsigned int>(num_mc_tasks))));
 		}
 		boost::timer::auto_cpu_timer t;
 //		knl.launch(s0, lig.content, lig.nv, lig.nf, lig.na, lig.np, seeds);
